factor out stuff from recipe detail activity to fragment
