+++ app/src/main/res/layout/activity_recipe_detail.xml
@@ -1,10 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent" android:layout_height="match_parent">
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/recipe_detail_recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-    </android.support.v7.widget.RecyclerView>
-
-</FrameLayout>
+<fragment xmlns:android="http://schemas.android.com/apk/res/android"
+    android:name="leonard.bakingapp.RecipeDetailFragment"
+    android:id="@+id/recipe_detail_fragment"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"/>
