+++ app/src/main/res/layout/fragment_recipe_detail.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.v7.widget.RecyclerView android:id="@+id/recipe_detail_recycler_view"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+</android.support.v7.widget.RecyclerView>
